moving common code
