+++ SwiftRuntimeLibrary.iOS/SwiftRuntimeLibrary.iOS.csproj
@@ -282,6 +282,12 @@
     <Compile Include="..\SwiftRuntimeLibrary\BaseAssociatedTypeProxy.cs">
       <Link>BaseAssociatedTypeProxy.cs</Link>
     </Compile>
+    <Compile Include="..\SwiftRuntimeLibrary\SwiftNativeObject.cs">
+      <Link>SwiftNativeObject.cs</Link>
+    </Compile>
+    <Compile Include="..\SwiftRuntimeLibrary\SwiftNativeInstance.cs">
+      <Link>SwiftNativeInstance.cs</Link>
+    </Compile>
   </ItemGroup>
   <Target Name="GeneratedCSCode" BeforeTargets="CoreCompile" Inputs="$(MSBuildProjectFullPath)" Outputs="GeneratedCode\BindingMetadata.iOS.cs">
     <Exec Command="make --directory=../SwiftRuntimeLibrary create-pinvokes-ios configuration=$(Configuration)" />
