+++ SwiftRuntimeLibrary/XamTrivialSwiftObject.cs
@@ -5,56 +5,21 @@ using System;
 using System.Runtime.InteropServices;
 
 namespace SwiftRuntimeLibrary {
-	public class XamTrivialSwiftObject : ISwiftObject {
-		bool disposed;
-
+	public class XamTrivialSwiftObject : SwiftNativeObject {
 		public XamTrivialSwiftObject ()
+			: base (NativeMethodsForXamTrivialSwiftObject.PIctor (NativeMethodsForXamTrivialSwiftObject.PImeta ()),
+				NativeMethodsForXamTrivialSwiftObject.PImeta (), SwiftObjectRegistry.Registry)
 		{
-			SwiftObject = NativeMethodsForXamTrivialSwiftObject.PIctor (NativeMethodsForXamTrivialSwiftObject.PImeta ());
-			SwiftCore.Retain (SwiftObject);
-			SwiftObjectRegistry.Registry.Add (this);
 		}
+
 		XamTrivialSwiftObject (IntPtr p, SwiftObjectRegistry registry)
+			: base (p, NativeMethodsForXamTrivialSwiftObject.PImeta (), registry)
 		{
-			SwiftObject = p;
-			SwiftCore.Retain (p);
-			registry.Add (this);
 		}
 
 		public static object XamarinFactory (IntPtr p)
 		{
 			return new XamTrivialSwiftObject (p, SwiftObjectRegistry.Registry);
-		}
-
-		public IntPtr SwiftObject { get; set; }
-
-		~XamTrivialSwiftObject ()
-		{
-			Dispose (false);
-		}
-		public void Dispose ()
-		{
-			Dispose (true);
-		}
-
-		protected virtual void Dispose (bool disposing)
-		{
-			if (!disposed) {
-				if (disposing) {
-					DisposeManagedResources ();
-				}
-				DisposeUnmanagedResources ();
-				disposed = true;
-			}
-		}
-
-		protected virtual void DisposeManagedResources ()
-		{
-		}
-
-		protected virtual void DisposeUnmanagedResources ()
-		{
-			SwiftCore.Release (SwiftObject);
 		}
 	}
